remove main + move fetchtoken to pkg + small fixes
