+++ tests/integration/api-gateway/gateway-tests/go.mod
@@ -3,7 +3,11 @@ module github.com/kyma-project/kyma/tests/integration/api-gateway/gateway-tests
 go 1.12
 
 require (
+	github.com/avast/retry-go v2.4.1+incompatible
 	github.com/golang/protobuf v1.3.2 // indirect
+	github.com/pkg/errors v0.8.1
+	github.com/stretchr/testify v1.3.0
+	golang.org/x/crypto v0.0.0-20190308221718-c2843e01d9a2 // indirect
 	golang.org/x/sys v0.0.0-20191010194322-b09406accb47 // indirect
 	google.golang.org/appengine v1.6.5 // indirect
 	k8s.io/apimachinery v0.0.0-20191006235458-f9f2f3f8ab02
