+++ tests/integration/api-gateway/gateway-tests/pkg/oauth2/types.go
